backend/backend-python/requirements.txt
# Core Data
pandas
numpy

# Database
pyodbc
sqlalchemy

# Google & AI
google-genai
google-api-python-client

# Profanity Filter
better-profanity

# Web
requests
fastapi
uvicorn[standard]

# Config
python-dotenv
azure-storage-blob

# Machine Learning
tensorflow==2.20.0
scikit-learn

# Utilities
python-dateutil
Pillow
openpyxl
reportlab

# Crypto
pycryptodome

# Document Parsing
pypdf
python-docx
